service: call through bash to workaround issues
In 35f4eb3 we removed the full path (/usr/bin/kmods-via-containers)
because we wanted to allow for the executable to be in either
/usr/local/bin (delivered by administrator) or /usr/bin (delivered
by the OS). Unfortunately this doesn't work on OSTree based systems
right now because /usr/local/ is actually under /var/usrlocal and is
not mounted at the time that systemd analyzes the units [1]. Systemd will
boot out this unit as bad because it can't find `kmods-via-containers`
in the path.

We can workaround this for now by just calling it via bash so that
systemd won't ignore our units and we can still use either /usr/bin/
or /usr/local/bin/.

[1] https://github.com/coreos/rpm-ostree/issues/1936